Allow clearing just contents of global areas
Former-commit-id: ace7f1509fc6751d9c69618aea44a5dc60f03329
Former-commit-id: 42ff960b4898ecf6ca362f4369d4d4a847df2269

+++ web/concrete/src/Workflow/Request/DeletePageRequest.php
@@ -63,7 +63,11 @@ class DeletePageRequest extends PageRequest
         $c = Page::getByID($this->getRequestedPageID());
         if ($c->getPageTypeHandle() == STACKS_PAGE_TYPE) {
             $c = Stack::getByID($this->getRequestedPageID());
-            $c->delete();
+            if ($c->getStackType() == Stack::ST_TYPE_GLOBAL_AREA && $c->isNeutralStack()) {
+                $c->cloneVersion(t('Global Area emptied'), true);
+            } else {
+                $c->delete();
+            }
             $wpr = new WorkflowProgressResponse();
             $wpr->setWorkflowProgressResponseURL(URL::to('/dashboard/blocks/stacks', 'stack_deleted'));
